requirements.txt
pymupdf==1.23.7
sentence-transformers==2.5.1
chromadb==0.4.22
fastapi==0.110.0
uvicorn==0.22.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.15.2
webdriver-manager==4.0.1

langchain-text-splitters==0.0.1
groq==0.14.0    
pypdf==4.2.0


pymongo==4.6.1
passlib[bcrypt]==4.0.1
python-jose==3.3.0
python-multipart==0.0.20